tmpl-module type added

+++ src/n2n/composer/module/ModuleInstaller.php
@@ -10,9 +10,19 @@ class ModuleInstaller extends LibraryInstaller {
 	 * @see \Composer\Installer\InstallerInterface::supports()
 	 */
 	public function supports($packageType) {
-		return $packageType == self::N2N_MODULE_TYPE;
+		return $packageType == self::N2N_MODULE_TYPE || $packageType == self::N2N_TMPL_MODULE_TYPE; 
 	}
 
+	
+	public function isInstalled(\Composer\Repository\InstalledRepositoryInterface $repo, \Composer\Package\PackageInterface $package) {
+		if (!$this->needsUpdate($package)) {
+			return true;
+		}
+
+		return parent::isInstalled($repo, $package);
+	}
+
+	
 	/**
 	 * {@inheritDoc}
 	 * @see \Composer\Installer\InstallerInterface::install()
@@ -22,6 +32,7 @@ class ModuleInstaller extends LibraryInstaller {
 		
 		$this->removeResources($package);
 		$this->installResources($package);
+		
 	}
 	/**
 	 * {@inheritDoc}
@@ -29,8 +40,9 @@ class ModuleInstaller extends LibraryInstaller {
 	 */
 	public function update(\Composer\Repository\InstalledRepositoryInterface $repo, \Composer\Package\PackageInterface $initial, 
 			\Composer\Package\PackageInterface $target) {
-
-		$this->moveBackResources($initial);
+		if (!$this->needsUpdate($initial)) return;
+				
+    	$this->moveBackResources($initial);
 				
 		parent::update($repo, $initial, $target);
 		
@@ -43,12 +55,23 @@ class ModuleInstaller extends LibraryInstaller {
 	 * @see \Composer\Installer\InstallerInterface::uninstall()
 	 */
 	public function uninstall(\Composer\Repository\InstalledRepositoryInterface $repo, \Composer\Package\PackageInterface $package) {
-		$this->moveBackResources($package);
+    	$this->moveBackResources($package);
+    	
+    	if (!$this->isTmplPackage($package)) {
+    		$pattern = '/' . $this->getModuleName($package);
+	    	$this->removeFromGitIgnore($this->getVarDestDirPath() . DIRECTORY_SEPARATOR . self::ETC_DIR,
+	    			$pattern);
+	    	$this->removeFromGitIgnore($this->getPublicDestDirPath() . DIRECTORY_SEPARATOR . self::ASSETS_DIR,
+	    			$pattern);
+    	}
 		
 		parent::uninstall($repo, $package);
 	}
 	
 	const N2N_MODULE_TYPE = 'n2n-module';
+	const N2N_TMPL_MODULE_TYPE = 'n2n-tmpl-module';
+	const APP_ORIG_DIR = 'src' . DIRECTORY_SEPARATOR . 'app';
+	const APP_DEST_DIR = '..' . DIRECTORY_SEPARATOR . 'app';
 	const VAR_ORIG_DIR = 'src' . DIRECTORY_SEPARATOR . 'var';
 	const VAR_DEST_DIR = '..' . DIRECTORY_SEPARATOR . 'var';
 	const ETC_DIR = 'etc';
@@ -58,6 +81,19 @@ class ModuleInstaller extends LibraryInstaller {
 	
 	private function getModuleName(Package $package) {
 		return pathinfo($this->getInstallPath($package), PATHINFO_BASENAME);
+	}
+	
+	private function getAppOrigDirPath(Package $package) {
+		return $this->filesystem->normalizePath($this->getInstallPath($package) . DIRECTORY_SEPARATOR 
+				. self::APP_ORIG_DIR);
+	}
+	
+	private function getAppDestDirPath() {
+		return $this->filesystem->normalizePath($this->vendorDir . DIRECTORY_SEPARATOR . self::APP_DEST_DIR);
+	}
+	
+	private function getRelAppDirPath(Package $package) {
+	    return DIRECTORY_SEPARATOR . str_replace('-', DIRECTORY_SEPARATOR, $this->getModuleName($package));
 	}
 	
 	private function getVarOrigDirPath(Package $package) {
@@ -87,9 +123,12 @@ class ModuleInstaller extends LibraryInstaller {
 	}
 	
 	public function moveBackResources(Package $package) {
+		if (!$this->needsUpdate($package)) return;
+		
 		$relEtcDirPath = $this->getRelEtcDirPath($package);
 		$mdlEtcOrigDirPath = $this->getVarOrigDirPath($package) . $relEtcDirPath;
 		$mdlEtcDestDirPath = $this->getVarDestDirPath() . $relEtcDirPath;
+		
 		if (is_dir($mdlEtcDestDirPath)) {
 			$this->filesystem->copyThenRemove($mdlEtcDestDirPath, $mdlEtcOrigDirPath);
 		}
@@ -103,10 +142,15 @@ class ModuleInstaller extends LibraryInstaller {
 	}
 	
 	private function removeResources(Package $package) {
+		if (!$this->needsUpdate($package)) return;
+		$moduleName = $this->getModuleName($package);
+		
 		$mdlEtcDestDirPath = $this->getVarDestDirPath() . $this->getRelEtcDirPath($package);
 		if (is_dir($mdlEtcDestDirPath)) {
 			try {
 				$this->filesystem->removeDirectory($mdlEtcDestDirPath);
+				$this->removeFromGitIgnore($this->getVarDestDirPath() . DIRECTORY_SEPARATOR . self::ETC_DIR, 
+						$moduleName);
 			} catch (\RuntimeException $e) {}
 		}
 		
@@ -114,16 +158,47 @@ class ModuleInstaller extends LibraryInstaller {
 		if (is_dir($mdlAssetsDestDirPath)) {
 			try {
 				$this->filesystem->removeDirectory($mdlAssetsDestDirPath);
+				$this->removeFromGitIgnore($this->getPublicDestDirPath() . DIRECTORY_SEPARATOR . self::ASSETS_DIR, 
+						$moduleName);
 			} catch (\RuntimeException $e) {}	
 		}
 	}
 	
 	private function installResources(Package $package) {
-		$this->moveEtc($package);
+		$this->moveApp($package);
 		$this->moveAssets($package);
+		$this->moveEtc($package);
+		
+		if ($this->isTmplPackage($package)) {
+			try {
+				$this->filesystem->removeDirectory($this->getInstallPath($package));
+			} catch (\RuntimeException $e) {}
+		}
+	}
+	
+	private function moveApp(Package $package) {
+		if (!$this->isTmplPackage($package) || $this->hasDestEtcDirPath($package)) return;
+		
+ 	    $appOrigDirPath = $this->getAppOrigDirPath($package);
+ 	    $appDestDirPath = $this->getAppDestDirPath();
+	    
+ 	    $this->valOrigDirPath($appOrigDirPath, $package);
+ 	    
+ 	    $relAppDirPath = $this->getRelAppDirPath($package);
+ 	    $mdlAppOrigDirPath = $appOrigDirPath . $relAppDirPath;
+ 	    $mdlAppDestDirPath = $appDestDirPath . $relAppDirPath;
+ 	    
+ 	    if (!is_dir($mdlAppOrigDirPath)) {
+ 	        return;
+ 	    }
+ 	    if (!is_dir($mdlAppDestDirPath) && $this->valDestDirPath($appDestDirPath, $package)) {
+	        $this->filesystem->copyThenRemove($appOrigDirPath, $appDestDirPath);
+	    }
 	}
 	
 	private function moveEtc(Package $package) {
+		if (!$this->needsUpdate($package)) return;
+		
 		$varOrigDirPath = $this->getVarOrigDirPath($package);
 		$varDestDirPath = $this->getVarDestDirPath();
 	
@@ -133,6 +208,7 @@ class ModuleInstaller extends LibraryInstaller {
 		$mdlEtcOrigDirPath = $varOrigDirPath . $relEtcDirPath;
 		$mdlEtcDestDirPath = $varDestDirPath . $relEtcDirPath;
 	
+		//don't move if etc folder exists and 
 		if (!is_dir($mdlEtcOrigDirPath)) {
 			return;
 		}
@@ -140,9 +216,16 @@ class ModuleInstaller extends LibraryInstaller {
 		if ($this->valDestDirPath($varDestDirPath, $package)) {
 			$this->filesystem->copyThenRemove($mdlEtcOrigDirPath, $mdlEtcDestDirPath);
 		}
+		
+		if (!$this->isTmplPackage($package)) {
+			$this->addToGitIgnore($varDestDirPath . DIRECTORY_SEPARATOR . self::ETC_DIR, 
+					'/' . $this->getModuleName($package));
+		}
 	}
 	
 	private function moveAssets(Package $package) {
+		if (!$this->needsUpdate($package)) return;
+		
 		$publicOrigDirPath = $this->getPublicOrigDirPath($package);
 		$publicDestDirPath = $this->getPublicDestDirPath();
 	
@@ -156,8 +239,13 @@ class ModuleInstaller extends LibraryInstaller {
 			return;
 		}
 	
-		if ($this->valDestDirPath($publicDestDirPath, $package)) {
+		if (!is_dir($mdlAssetsDestDirPath) && $this->valDestDirPath($publicDestDirPath, $package)) {
 			$this->filesystem->copyThenRemove($mdlAssetsOrigDirPath, $mdlAssetsDestDirPath);
+		}
+		
+		if (!$this->isTmplPackage($package)) {
+			$this->addToGitIgnore($publicDestDirPath . DIRECTORY_SEPARATOR . self::ASSETS_DIR, 
+					'/' . $this->getModuleName($package));
 		}
 	}
 	
@@ -165,7 +253,7 @@ class ModuleInstaller extends LibraryInstaller {
 		if (is_dir($origDirPath)) return;
 	
 		$dirName = pathinfo($origDirPath, PATHINFO_BASENAME);
-		throw new CorruptedN2nModuleException($package->getPrettyName() . ' has type \'' . self::N2N_MODULE_TYPE
+		throw new CorruptedN2nModuleException($package->getPrettyName() . ' has type \'' . $package->getType()
 				. '\' but contains no ' . $dirName . ' directory: ' . $origDirPath);
 	}
 	
@@ -174,13 +262,65 @@ class ModuleInstaller extends LibraryInstaller {
 	
 		$dirName = pathinfo($destDirPath, PATHINFO_BASENAME);
 	
-		$question = $package->getPrettyName() . ' is an ' . self::N2N_MODULE_TYPE
+		$question = $package->getPrettyName() . ' is an ' . $package->getType()
 				. ' and requires a ' . $dirName . ' directory (' . $destDirPath
 				. '). Do you want to skip the installation of the ' . $dirName . ' files? [y,n] (default: y): ';
 		if ($this->io->askConfirmation($question)) return false;
 	
-		throw new N2nModuleInstallationException('Failed to install ' . self::N2N_MODULE_TYPE . ' '
+		throw new N2nModuleInstallationException('Failed to install ' . $package->getType() . ' '
 				. $package->getPrettyName() . '. Reason: ' . $dirName . ' directory missing: ' . $destDirPath);
+	}
+	
+	private function needsUpdate(Package $package) {
+		if (!$this->isTmplPackage($package)) return true;
+		
+		return !$this->hasDestEtcDirPath($package);
+	}
+	
+	private function hasDestEtcDirPath(Package $package) {
+		return is_dir($this->getVarDestDirPath() . $this->getRelEtcDirPath($package));
+	}
+	
+	private function isTmplPackage(Package $package) {
+		return $package->getType() === self::N2N_TMPL_MODULE_TYPE;
+	}
+	
+	private function addToGitIgnore($dirPath, $pattern) {
+		if (!is_dir($dirPath)) return;
+		
+		$filename = $dirPath . DIRECTORY_SEPARATOR . '.gitignore';
+		
+		$contents = [];
+		if (is_file($filename)) {
+			$contents = file($filename);
+		}
+		
+		foreach ($contents as $content) {
+			if (trim($content) == $pattern) return;
+		}
+		
+		$contents[] = (!empty($contents) ? PHP_EOL : '') . $pattern;
+		
+		
+		file_put_contents($filename, $contents);
+	}
+	
+	private function removeFromGitIgnore($dirPath, $pattern) {
+		if (!is_dir($dirPath)) return;
+		
+		$filename = $dirPath . DIRECTORY_SEPARATOR . '.gitignore';
+		
+		$contents = [];
+		if (is_file($filename)) return;
+		
+		$newContents = [];
+		foreach ($contents as $content) {
+			if (trim($content) == $pattern) continue;
+			
+			$newContents[] = $content;
+		}
+		
+		file_put_contents($filename, $contents);
 	}
 	
 // 	private function copy($source, $target) {
